. seqtgmgr.h: fixed two bugs, one is that the task for RemoveTg was not marked the Pending flag and caused the client notify task not called and the Unmount process hung. The other is that RemoveTg could be called after the parent object is gone, and result in segment fault. And the solution is to hold a reference to the parent object.
